requirements.txt
PyGithub
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-groq
langchain-text-splitters
sentence-transformers
chromadb
fastapi[standard]
groq
sqlalchemy